chore(debugging): add probe registry (#3922)
## Description

The probe registry is used to keep track of all the probes that are received, instrumented or removed. Since this is a potentially shared resource across threads, modifications are guarded by a re-entrant lock. The registry also uses an instance of the probe status logger to report probe status in a centralised way.

### Typical Probe Life-cycle

In general terms, when a new probe is received from the configuration API, it should be registered with the probe registry. This will emit a `RECEIVED` state message to signal to the backend that the probe was received. If it also has valid location information, it is also marked for further processing on a location-based way (for performance). Here, further processing is the action performed, e.g., by import hooks to instrument the code according to the information carried by the probe object. It is responsibility of these hooks to then call `set_installed` from the registry on the probe to mark it as `INSTALLED`. Pending probes can be queried from the registry on a location basis with the `get_pending` method. The rationale is that import hooks will want to act on probes that are related to the module the hooks are called on, which correlates with probe's location information.

When a probe is no longer required, it can be removed with `
unregister`. This ensures that probes that are pending are also removed, and not processed by accident.

## Checklist
- [x] Title must conform to [conventional commit](https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional).
- [x] Add additional sections for `feat` and `fix` pull requests.
- [ ] Ensure tests are passing for affected code.
- [x] [Library documentation](https://github.com/DataDog/dd-trace-py/tree/1.x/docs) and/or [Datadog's documentation site](https://github.com/DataDog/documentation/) is updated. Link to doc PR in description.

## Reviewer Checklist
- [ ] Title is accurate.
- [ ] Description motivates each change.
- [ ] No unnecessary changes were introduced in this PR.
- [ ] PR cannot be broken up into smaller PRs.
- [ ] Avoid breaking [API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces) changes unless absolutely necessary.
- [ ] Tests provided or description of manual testing performed is included in the code or PR.
- [x] Release note has been added for fixes and features, or else `changelog/no-changelog` label added.
- [ ] All relevant GitHub issues are correctly linked.
- [ ] Backports are identified and tagged with Mergifyio.
- [ ] Add to milestone.