fix(internal): span sampling rule match logic (#3992)
## Description
Previous logic did not cover the possibility of spans not having a service name or not having an operation name.  Previously if the span only had a service or only had and operation name, we would get an error when the GlobMatcher tried to take the length of the subject, which for whichever was missing (the name or the service) was None `len(None)`: https://github.com/DataDog/dd-trace-py/blob/1.x/ddtrace/internal/glob_matching.py#L24

## Testing strategy
I've added tests for the span sampling rule to test matching on spans with only a service name and spans with only an operation name.

## Checklist
- [ ] Title must conform to [conventional commit](https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional).
- [ ] Add additional sections for `feat` and `fix` pull requests.
- [ ] Ensure tests are passing for affected code.
- [ ] [Library documentation](https://github.com/DataDog/dd-trace-py/tree/1.x/docs) and/or [Datadog's documentation site](https://github.com/DataDog/documentation/) is updated. Link to doc PR in description.


## Reviewer Checklist
- [ ] Title is accurate.
- [ ] Description motivates each change.
- [ ] No unnecessary changes were introduced in this PR.
- [ ] PR cannot be broken up into smaller PRs.
- [ ] Avoid breaking [API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces) changes unless absolutely necessary.
- [ ] Tests provided or description of manual testing performed is included in the code or PR.
- [ ] Release note has been added for fixes and features, or else `changelog/no-changelog` label added.
- [ ] All relevant GitHub issues are correctly linked.
- [ ] Backports are identified and tagged with Mergifyio.
- [ ] Add to milestone.