fix(pytest): fix encoding of some objects in pytest test parameters (#4014)
## Description
We use `repr` to get the string representation of test parameter values in `pytest`. This normally works but doesn't for some values, such as functions: 

```python
def my_func():
    print('do stuff')

b = iter([1,2])

print(repr(b))
print(repr(my_func))
```

This results in 
```
<list_iterator object at 0x101119fa0>
<function my_func at 0x10107fee0>
```

This is because `repr` relies on [id](https://docs.python.org/3/library/functions.html#id) to get the string representation of some objects. 

We rely on the test parameters to calculate a test fingerprint, but if the test parameters change across executions, the calculation falls apart. 

This PR removes the `at 0x101119fa0` of the string to get a more stable representation, so the above example would look like: 

```
<list_iterator object>
<function my_func>
```


## Checklist
- [x] Title must conform to [conventional commit](https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional).
- [ ] Add additional sections for `feat` and `fix` pull requests.
- [x] Ensure tests are passing for affected code.
- [ ] [Library documentation](https://github.com/DataDog/dd-trace-py/tree/1.x/docs) and/or [Datadog's documentation site](https://github.com/DataDog/documentation/) is updated. Link to doc PR in description.





## Motivation


## Design 


## Testing strategy




## Relevant issue(s)


## Testing strategy




## Reviewer Checklist
- [x] Title is accurate.
- [x] Description motivates each change.
- [x] No unnecessary changes were introduced in this PR.
- [x] PR cannot be broken up into smaller PRs.
- [x] Avoid breaking [API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces) changes unless absolutely necessary.
- [x] Tests provided or description of manual testing performed is included in the code or PR.
- [x] Release note has been added for fixes and features, or else `changelog/no-changelog` label added.
- [ ] All relevant GitHub issues are correctly linked.
- [ ] Backports are identified and tagged with Mergifyio.
- [ ] Add to milestone.