feat(internal): environment config for span sampling (#3977)
## Description
In order for users to configure the single span sampling rules, this introduces the ability to do that via the environment variable `DD_SPAN_SAMPLING_RULES`. This PR also initializes the single span sampling processor if rules are provided.

## Motivation
This is required in order for users to specify the rules they would like to be applied for single span sampling.

## Design 
Most of the work introduced in this PR is done in sampling.py which is essentially responsible for: 
1. Pulling in and validating the rules from the environment `DD_SPAN_SAMPLING_RULES`
2. Providing those rules to the span processor so it can apply them to spans

## Testing strategy
Testing mostly revolves around making sure that different configurations are picked up correctly, default values are what we expect, and validation works as intended. 

In addition some tests were done around making sure that spans are actually sampled as expected, however a great deal of that has already been done in https://github.com/DataDog/dd-trace-py/pull/3966, and will be done in a future testing PR(which will essentially validate E2E) so it seemed unnecessary to go much further.

## Checklist
- [x] Title must conform to [conventional commit](https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional).
- [x] Add additional sections for `feat` and `fix` pull requests.
- [x] Ensure tests are passing for affected code.
- [x] [Library documentation](https://github.com/DataDog/dd-trace-py/tree/1.x/docs) and/or [Datadog's documentation site](https://github.com/DataDog/documentation/) is updated. Link to doc PR in description.

## Reviewer Checklist
- [x] Title is accurate.
- [x] Description motivates each change.
- [x] No unnecessary changes were introduced in this PR.
- [x] PR cannot be broken up into smaller PRs.
- [x] Avoid breaking [API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces) changes unless absolutely necessary.
- [x] Tests provided or description of manual testing performed is included in the code or PR.
- [x] Release note has been added for fixes and features, or else `changelog/no-changelog` label added.
- [ ] All relevant GitHub issues are correctly linked.
- [ ] Backports are identified and tagged with Mergifyio.
- [ ] Add to milestone.