feat(snowflake): add snowflake query id tag (#4069)
## Description
We are using snowflake as our database and wanted to add a tag that represents the "query id" so we will be able to look for it later in snowflake.

For example, the query id in snowflake is called "sfqid" and returned after the query executed.
So in this example, I've added "sfqid" with value None to the Pin inside snowflake/patch.py, and then after the query execution I'm populating the value of this tag (inside dbapi/__init__.py)


![image](https://user-images.githubusercontent.com/12069200/183624054-5cfed3fa-9eae-4807-9b22-ef2de00982d9.png)

Now I can take this sfqid and look for it in the Database 
![image](https://user-images.githubusercontent.com/12069200/183624153-db3f3acc-f757-418b-8d8b-c65c1d240a57.png)



## Checklist
- [ ] Title must conform to [conventional commit](https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional).
- [ ] Add additional sections for `feat` and `fix` pull requests.
- [ ] Ensure tests are passing for affected code.
- [ ] [Library documentation](https://github.com/DataDog/dd-trace-py/tree/1.x/docs) and/or [Datadog's documentation site](https://github.com/DataDog/documentation/) is updated. Link to doc PR in description.





## Motivation
Add the ability to "inject" tags that are available only after the query done (like rows_count).

## Design 
I'm using the fact that Pin is a dictionary of tags that being passed to the dbapi, to inject tags with None value that will later be populated.

## Testing strategy
The feature is working and currently snowflake spans gets "sfqid"'s.




## Reviewer Checklist
- [ ] Title is accurate.
- [ ] Description motivates each change.
- [ ] No unnecessary changes were introduced in this PR.
- [ ] PR cannot be broken up into smaller PRs.
- [ ] Avoid breaking [API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces) changes unless absolutely necessary.
- [ ] Tests provided or description of manual testing performed is included in the code or PR.
- [ ] Release note has been added for fixes and features, or else `changelog/no-changelog` label added.
- [ ] All relevant GitHub issues are correctly linked.
- [ ] Backports are identified and tagged with Mergifyio.
- [ ] Add to milestone.