ci(debugging): add exploration testing (#3991)
This change adds the so-called exploration tests which we use to ensure that the dynamic instrumentation does not affect the expected behaviour of the instrumented code. We select a few popular Python frameworks and inject probes on each single line and wrap every single function. We then run the full test suite for the actual validation.

At the end of the test run, line coverage and function call counts are reported as a "proof-of-work"