fix: use soft forksafe hooks with start-up threads (#4376)
Services like RCM and Dynamic Instrumentation (which depends on the former) are likely to start worker threads on start-up when enabled. This is known to cause issues with some frameworks if ordinary fork hooks are used. This change ensures that we use "soft" hooks, that is hooks invoked after the os.fork call returns, by default to mitigate the problem.