chore(ci/profiling): increase profiling windows jobs parallelism (#4573)
## Description

This improves the runtime of the profiling windows jobs from 20+m each to around 6m each.

Each profiling windows job is broken up by python version, so we can only set the parallelism to the max number of tox environments for each. For older Python's this is 4, and newer this is 5. This means each tox env will run on it's own node.

## Reviewer Checklist
- [x] Title is accurate.
- [x] Description motivates each change.
- [x] No unnecessary changes were introduced in this PR.
- [x] Avoid breaking [API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces) changes unless absolutely necessary.
- [x] Tests provided or description of manual testing performed is included in the code or PR.
- [x] Release note has been added for fixes and features, or else `changelog/no-changelog` label added.
- [x] All relevant GitHub issues are correctly linked.
- [x] Backports are identified and tagged with Mergifyio.