ci: reduce number of spans in stats snapshot tests (#4633)
Having a large number of spans could cause multiple stats payloads to be submitted while the test case is running. This is made apparent by the test cases that generate many spans being flakier than the ones that do not.

An alternative solution could be to change the flush interval to a very large value to prevent multiple buckets from being emitted. However, this approach would change the configuration to one that users wouldn't likely use.


## Reviewer Checklist
- [x] Title is accurate.
- [x] Description motivates each change.
- [x] No unnecessary changes were introduced in this PR.
- [x] Avoid breaking [API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces) changes unless absolutely necessary.
- [x] Tests provided or description of manual testing performed is included in the code or PR.
- [x] Release note has been added and follows the [library release note guidelines](https://ddtrace.readthedocs.io/en/stable/contributing.html#Release-Note-Guidelines), or else `changelog/no-changelog` label added.
- [x] All relevant GitHub issues are correctly linked.
- [x] Backports are identified and tagged with Mergifyio.