refactor(stack): do not try to account CPU time on gevent tasks (#4474)
The current stack collector maps the CPU usage of a thread to its main
greenlet. This means that the following happens:
```
  MainThread +-> main gevent task (CPU + wall time)
             |-> other gevent task 1 (wall time)
             `-> other gevent task 2 (wall time)
```
The issue with this approach is that when using an application like
gunicorn,
the end result is suboptimal: the main gevent task appears to be
sleeping in
gunicorn main loop (because it's waiting for I/O) and hides potential
real code
running in the main thread.

The original idea was to hide the MainThread frames completely as they
represent the gevent hub running most of the time, and replace it with
the
gevent main task. However, it turns out that with software like
gunicorn, this
loses information.

This new approach does not try to be smarter, and take a step back. It
will now
accounts the CPU time on the MainThread, and list all the tasks only for
the
wall time:
```
  MainThread (CPU + wall time) +-> main gevent task (wall time)
                               |-> other gevent task 1 (wall time)
                               `-> other gevent task 2 (wall time)
```
This does not try anymore to assign the CPU time to any gevent task, but
keeps
the CPU time on the main thread. However, it shows the details of the
gevent
tasks running in the wall time, making it clear for a user what's
running.

As there is no way to account for the CPU time of the different gevent
task
(they run in the same thread), this is a good representation of the
reality
anyway.

This also has the upside to bring gevent tasks representation closer to
what is
done for asyncio.

Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>