chore(logging): fix noisy rq patching log (#4779)
The patching logs generated a scary looking message for rq:

```
DEBUG:ddtrace._monkey:failed to patch rq
Traceback (most recent call last):
  File "/Users/kverhoog/dev/dd-trace-py/ddtrace/_monkey.py", line 249, in _patch_module
    return _attempt_patch_module(module, patch_modules_prefix=patch_modules_prefix)
  File "/Users/kverhoog/dev/dd-trace-py/ddtrace/_monkey.py", line 304, in _attempt_patch_module
    imported_module.patch()
  File "/Users/kverhoog/dev/dd-trace-py/ddtrace/contrib/rq/__init__.py", line 207, in patch
    import rq
ModuleNotFoundError: No module named 'rq'
```

because the rq integration imports the `rq` module on `patch()` and
doesn't declare required_modules. Nor does it dynamically expose the
patch and unpatch methods.

By making the rq integration a patch-on-import integration the issue is
avoided. This can be done because the rq integration was written to be
patch-on-import compatible (no top-level imports of the library).