fix(lib-injection): publish release tag of image (#4931)
Currently, only dev images for commits are pushed to the container
registry. This means that images like `v1.7.0` that get published are
really just aliases of the corresponding SHA image. These images use

```
pip install git+....@sha
```
to install the image.

For example:

```
~/d/h/c/datadog ❯❯❯ docker run -it gcr.io/datadoghq/dd-lib-python-init:v1.7.0 cat /datadog-init/sitecustomize.py | head -n 12
"""
This module when included on the PYTHONPATH will install the ddtrace package from pypi
for the Python runtime being used.
"""
$ docker run -it gcr.io/datadoghq/dd-lib-python-init:v1.7.0 cat /datadog-init/sitecustomize.py | head -n 12
...
version = "git+https://github.com/Datadog/dd-trace-py@ced65df45a9b592dd36af12c37c047c1486af38a"
```

While this isn't _wrong_ it is slow for auto instrumented applications
to start since the install is from source. It also means all
installation dependencies are required and the install could fail.

The solution is to run the publishing Github workflow on release after
the pypi_upload job. This should ensure that we build an image that uses
the release tag and subsequently pip installs with that version rather
than from source.

The Gitlab job responsible for publishing the image to the other
container registries is updated to use the release tag image rather than
the SHA. There is and was a race between when the Github registry image
is published and when the Gitlab job runs. To mitigate this the Gitlab
job is delayed by 1 day to give time for the image to be pushed to the
Github registry.