chore(rq): wait for snapshot traces in flaky rq test (#4986)
The rq test `test_enqueue()` has been found to be flaky in CI runs
(example
[run](https://app.circleci.com/pipelines/github/DataDog/dd-trace-py/27434/workflows/9c729ac6-9ee7-4c25-a9ac-b90fba66f419/jobs/1873546)),
with traces not immediately being found. The solution is to use the
snapshot function's `wait_for_num_traces` option to make the test retry
fetching traces from the testagent for a few seconds.

Note: this workaround is avoided in Python 3.5 as for some reason the
snapshot function's `wait_for_num_traces` retry mechanism does not work
in 3.5.