chore(profiling): string table code reuse (#4941)
This change improves code reuse in the profiler by substituting another
implementation with the one in the internal area of the library. This
also has the extra benefit of avoiding a sorting step, thus providing
slightly better performance.

A custom sequence object has also been removed in favour of
`itertools.count`.